13-20 changes.
Rite of The Dead, Rite of The Oracle, and Rite of the Roar. Includes Predatory Rites version. 
Added Hardened Soul
Added Sanguine Mastery (kind of) the restore a blood maledict on crit works fine. Not sure about reduced rite damage.
Added Blood Curse of the Exorcist. (Did not add amplified version.)

Tried to add Rite Revival, but was unable to get it to work. Used Relentless Endurance as base, if you wanted to take a stab at it.